+++ R/description.R
@@ -44,11 +44,12 @@ pac_description <- function(pac,
     version <- if (!version_null) {
       version
     } else if (!is.null(at)) {
-      vv <- utils::tail(pac_timemachine(pac, at = at)$Version, 1)
-      if (isNA(vv) || is.null(vv)) {
+      res <- pac_timemachine(pac, at = at)
+      if (isNA(res) || is.null(res)) {
         return(NA)
+      } else {
+        utils::tail(res$Version, 1)
       }
-      vv
     } else {
       last_version
     }