+++ NEWS.md
@@ -1,3 +1,9 @@
+# pacs 0.4.10.9000
+
+* protect against 404/403 results from crandb, linked with https://github.com/r-hub/pkgsearch/issues/117#issuecomment-1651081139
+* unstable crandb causes that lifeduration and timemachine functions use by default cran direct calls.
+* add informative messages when a web call failed.
+
 # pacs 0.4.10
 
 * remove the not needed testthat log file.